+++ Card/V1/index.spec.ts
@@ -1,15 +1,17 @@
-import * as model from "../index"
+import * as model from "../../index"
 
 describe("Card", () => {
 	it("is", async () => {
-		const card: model.Card = {
+		const card: model.Card.V1 = {
+			id: "1234abcd",
+			reference: "abc0123",
 			scheme: "visa",
 			iin: "510510",
 			last4: "9000",
 			expires: [1, 21],
 			type: "credit",
 		}
-		expect(model.Card.is(card)).toBeTruthy()
+		expect(model.Card.V1.is(card)).toBeTruthy()
 	})
 	it("from", async () => {
 		const card: model.Card.Creatable = {
@@ -17,7 +19,7 @@ describe("Card", () => {
 			expires: [2, 22],
 			csc: "123",
 		}
-		const complete: model.Card = model.Card.from(card)
+		const complete: Omit<model.Card.V1, "id" | "reference"> = model.Card.V1.from(card)
 		expect(complete).toEqual({
 			scheme: "mastercard",
 			iin: "510510",
@@ -27,7 +29,7 @@ describe("Card", () => {
 		const partialCard: model.Card.Change = {
 			pan: "5105105105105100",
 		}
-		const partial: Partial<model.Card> = model.Card.from(partialCard)
+		const partial: Omit<Partial<model.Card.V1>, "id" | "reference"> = model.Card.V1.from(partialCard)
 		expect(partial).toEqual({
 			scheme: "mastercard",
 			iin: "510510",
@@ -36,12 +38,12 @@ describe("Card", () => {
 		const expiresOnly: model.Card.Change = {
 			expires: [2, 22],
 		}
-		const expiresOnlyOutput: Partial<model.Card> = model.Card.from(expiresOnly)
+		const expiresOnlyOutput: Omit<Partial<model.Card.V1>, "id" | "reference"> = model.Card.V1.from(expiresOnly)
 		expect(expiresOnlyOutput).toEqual({
 			expires: [2, 22],
 		})
 		const emptyCard: model.Card.Change = {}
-		const emptyCardOutput: Partial<model.Card> = model.Card.from(emptyCard)
+		const emptyCardOutput: Omit<Partial<model.Card.V1>, "id" | "reference"> = model.Card.V1.from(emptyCard)
 		expect(emptyCardOutput).toEqual({})
 	})
 })
